Add files via upload
Shortly after the 2016 election, congress released over 2,600 Facebook ads and the actual advertisements themselves as PDFs. I wrote a python script to extract the data from the PDF and converted it into a csv. I also published a paper on the data here, which you can check out here, but please don't distract yourself on that right now, you have an exam to finish.

Each row is an advertisement that ran. The columns correspond to the attributes of the ad: the targeting parameters, when it ran, the amount spent and so on.

For this dataset, you're going to try and predict the "Ad Clicks", that is, the amount of clicks an ad got.