Add files via upload
The US government implemented the Paycheck Protection Program (PPP) as a crucial measure to provide support to small businesses impacted by the COVID-19 pandemic. This report aims to analyze the PPP loan data available on Kaggle, which includes detailed information on the loans approved under the program, to derive insights into the distribution of loans across various businesses, the industries that received the highest amount of funding, and the correlation between loan amounts and job retention. By analyzing this data, we can gain a better understanding of the program's effectiveness and how it impacted small businesses across different sectors and regions.